Moved matching DTO into matching interface.
