+++ jgrapht-core/src/main/java/org/jgrapht/alg/KuhnMunkresMinimalWeightBipartitePerfectMatching.java
@@ -134,21 +134,7 @@ public class KuhnMunkresMinimalWeightBipartitePerfectMatching<V, E>
         for (E e : m) {
             w += graph.getEdgeWeight(e);
         }
-        final double weight = w;
-        return new Matching<E>()
-        {
-            @Override
-            public double getWeight()
-            {
-                return weight;
-            }
-
-            @Override
-            public Set<E> getEdges()
-            {
-                return m;
-            }
-        };
+        return new MatchingImpl<E>(m, w);
     }
 
     /**
