+++ jgrapht-core/src/main/java/org/jgrapht/alg/MaximumWeightBipartiteMatching.java
@@ -110,21 +110,7 @@ public class MaximumWeightBipartiteMatching<V, E>
         for (E e : m) {
             w += graph.getEdgeWeight(e);
         }
-        final double weight = w;
-        return new Matching<E>()
-        {
-            @Override
-            public double getWeight()
-            {
-                return weight;
-            }
-
-            @Override
-            public Set<E> getEdges()
-            {
-                return m;
-            }
-        };
+        return new MatchingImpl<E>(m, w);
     }
 
     private void initializeVerticesAndEdges()
